requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
ibm-cloud-sdk-core==3.23.0
ibm-watson==9.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.0.2
openai==0.28.0
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
